feat: Add top-k evaluator, and new models (LSA, LDA, TFIDF) (#83)
* Created new ES reranker class with topk() method, created new evaluator scrip

* Add LSA, tfidf, LDA distance ranking

* Remove duplicate `build_isolator`

* Subclass ElasticSearchReRanker with ElasticSearchTopK

* Delete "enhanced" ES

Too much duplicate. this is replaced with the new class subclassing ElasticSearchReRanker

* Add option to evaluator.py to perform top-k

we can avoid duplicating the script - and we can run all in one shot..

* Delete top_k_evaluator.py

we added an option in evaluator.py to peform top-k as well. so this avoids duplicating the script - and we can run all in one shot..

* Add readme to covidfaq.evaluating

* linting

Co-authored-by: lp-noel <25254389+lp-noel@users.noreply.github.com>